+++ build_web_compilers/lib/src/dev_compiler_bootstrap.dart
@@ -37,6 +37,7 @@ Future<void> bootstrapDdc(
   String entrypointExtension = jsEntrypointExtension,
   required bool? nativeNullAssertions,
   bool usesWebHotReload = false,
+  bool ddcLibraryBundle = false,
   bool unsafeAllowUnsupportedModules = false,
 }) async {
   // Ensures that the sdk resources are built and available.
@@ -112,7 +113,7 @@ $librariesString
   final dartEntrypointParts = _context.split(dartEntrypointId.path);
   final packageName = module.primarySource.package;
   final entrypointLibraryName =
-      usesWebHotReload
+      ddcLibraryBundle
           ? _context.joinAll([
             // Convert to a package: uri for files under lib.
             if (dartEntrypointParts.first == 'lib') 'package:$packageName',
@@ -132,7 +133,7 @@ $librariesString
   String entrypointJsContent;
   String bootstrapContent;
   String bootstrapEndContent;
-  if (usesWebHotReload) {
+  if (ddcLibraryBundle) {
     final ddcSdkUrl =
         r'packages/build_web_compilers/src/dev_compiler/dart_sdk.js';
     modulePaths['dart_sdk'] = ddcSdkUrl;
@@ -148,8 +149,8 @@ $librariesString
               : _context.joinAll(_context.split(jsId.path).skip(1));
     }
     final bootstrapEndModuleName = _context.relative(
-      bootstrapId.path,
-      from: _context.dirname(bootstrapEndId.path),
+      bootstrapEndId.path,
+      from: _context.dirname(bootstrapId.path),
     );
     bootstrapContent = generateDDCLibraryBundleMainModule(
       entrypoint: entrypointLibraryName,