+++ build_web_compilers/lib/src/dev_compiler_builder.dart
@@ -42,8 +42,8 @@ class DevCompilerBuilder implements Builder {
   /// Enables canary features in DDC.
   final bool canaryFeatures;
 
-  /// Emits code with the DDC module system.
-  final bool ddcModules;
+  /// Emits code with the DDC Library Bundle module system.
+  final bool ddcLibraryBundle;
 
   final bool trackUnusedInputs;
 
@@ -74,7 +74,7 @@ class DevCompilerBuilder implements Builder {
     this.generateFullDill = false,
     this.emitDebugSymbols = false,
     this.canaryFeatures = false,
-    this.ddcModules = false,
+    this.ddcLibraryBundle = false,
     this.trackUnusedInputs = false,
     required this.platform,
     String? sdkKernelPath,
@@ -125,16 +125,16 @@ class DevCompilerBuilder implements Builder {
       await _createDevCompilerModule(
         module,
         buildStep,
-        useIncrementalCompiler,
-        generateFullDill,
-        emitDebugSymbols,
-        canaryFeatures,
-        ddcModules,
-        trackUnusedInputs,
-        platformSdk,
-        sdkKernelPath,
-        librariesPath,
         environment,
+        useIncrementalCompiler: useIncrementalCompiler,
+        generateFullDill: generateFullDill,
+        emitDebugSymbols: emitDebugSymbols,
+        canaryFeatures: canaryFeatures,
+        ddcLibraryBundle: ddcLibraryBundle,
+        trackUnusedInputs: trackUnusedInputs,
+        dartSdk: platformSdk,
+        sdkKernelPath: sdkKernelPath,
+        librariesPath: librariesPath,
       );
     } on DartDevcCompilationException catch (e) {
       await handleError(e);
@@ -148,16 +148,16 @@ class DevCompilerBuilder implements Builder {
 Future<void> _createDevCompilerModule(
   Module module,
   BuildStep buildStep,
-  bool useIncrementalCompiler,
-  bool generateFullDill,
-  bool emitDebugSymbols,
-  bool canaryFeatures,
-  bool ddcModules,
-  bool trackUnusedInputs,
-  String dartSdk,
-  String sdkKernelPath,
-  String librariesPath,
   Map<String, String> environment, {
+  required bool useIncrementalCompiler,
+  required bool generateFullDill,
+  required bool emitDebugSymbols,
+  required bool canaryFeatures,
+  required bool ddcLibraryBundle,
+  required bool trackUnusedInputs,
+  required String dartSdk,
+  required String sdkKernelPath,
+  required String librariesPath,
   bool debugMode = true,
 }) async {
   final transitiveDeps = await buildStep.trackStage(
@@ -202,11 +202,11 @@ Future<void> _createDevCompilerModule(
       WorkRequest()
         ..arguments.addAll([
           '--dart-sdk-summary=$sdkSummary',
-          '--modules=${ddcModules ? 'ddc' : 'amd'}',
+          '--modules=${ddcLibraryBundle ? 'ddc' : 'amd'}',
           '--no-summarize',
           if (generateFullDill) '--experimental-output-compiled-kernel',
           if (emitDebugSymbols) '--emit-debug-symbols',
-          if (canaryFeatures) '--canary',
+          if (canaryFeatures || ddcLibraryBundle) '--canary',
           '-o',
           jsOutputFile.path,
           debugMode ? '--source-map' : '--no-source-map',