+++ build_modules/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.1.5
+
+- Add support for DDC's Library Bundle module system, which is compatible with web hot reload. This is not yet enabled by default.
+
 ## 5.1.4
 
 - Fix module_builder reading DDC modules for non-primary dart libraries.